refactor - separate components & add AsyncBoundary
