+++ src/components/Product/organisms/ProductDetail/style.module.scss
@@ -1,0 +1,3 @@
+.shop-detail-info {
+  padding: 16px 23px 20px;
+}
