+++ src/pages/shop/[id]/index.tsx
@@ -1,23 +1,19 @@
 import { GetServerSidePropsContext } from 'next';
 
-import { useCallback } from 'react';
+import { ReactElement } from 'react';
 
+import ErrorFallback from '@atoms/ErrorFallback';
 import HeadMeta from '@atoms/HeadMeta';
+import Loading from '@atoms/Loading';
 import { queryKey } from '@constants/react-query';
 import { seoData } from '@constants/seo';
 import { dehydrate, QueryClient } from '@tanstack/react-query';
+import AsyncBoundary from '@templates/AsyncBoundary';
 import Layout from '@templates/Layout';
 import { withGetServerSideProps } from 'src/api/core/withGetServerSideProps';
 import { getProductDetail } from 'src/api/product';
-import SellerComment from 'src/components/Product/molecules/SellerComment';
-import ProductBasic from 'src/components/Product/organisms/ProductBasic';
-import ProductFooter from 'src/components/Product/organisms/ProductFooter';
-import ProductImgSlide from 'src/components/Product/organisms/ProductImgSlide';
-import ProductNotice from 'src/components/Product/organisms/ProductNotice';
-import ProductSize from 'src/components/Product/organisms/ProductSize';
-import ProfileInfo from 'src/components/Product/organisms/ProfileInfo';
-import { useProdutDetail } from 'src/hooks/api/product';
-import { useSearchStore } from 'src/store/useSearchStore';
+import ProductDetail from 'src/components/Product/organisms/ProductDetail';
+import ProductRecommend from 'src/components/Product/organisms/ProductRecommend';
 
 import $ from './style.module.scss';
 
@@ -41,58 +37,36 @@ export const getServerSideProps = withGetServerSideProps(
 );
 
 function ShopDetail({ id }: { id: string }) {
-  const { data } = useProdutDetail(id);
-  const addProduct = useSearchStore(
-    useCallback((state) => state.addProduct, []),
+  return (
+    <>
+      <HeadMeta
+        title="re:Fashion | 상품 상세보기"
+        url={`${seoData.url}/shop/${id}`}
+      />
+
+      <AsyncBoundary
+        suspenseFallback={<Loading />}
+        errorFallback={ErrorFallback}
+      >
+        <ProductDetail {...{ id }} />
+      </AsyncBoundary>
+
+      <AsyncBoundary
+        suspenseFallback={<Loading />}
+        errorFallback={ErrorFallback}
+      >
+        <ProductRecommend {...{ id }} />
+      </AsyncBoundary>
+    </>
   );
-  const detailData = data?.data;
-
-  if (detailData) {
-    const { isMe, isSoldOut, sellerInfo, basic, sellerNotice } = detailData;
-    const { measure, opinion, price, isIncludeDelivery } = detailData;
-    const { updatedAt, like, view, contact } = detailData;
-    const { userId, profileImg: profileImage, nickname: name } = sellerInfo;
-    // const status = 'soldout'; // TODO: 백엔드와 협의, 추후에 상품 상태 추가
-    addProduct({ id: +id, img: sellerInfo.image[0] });
-
-    return (
-      <>
-        <HeadMeta
-          title="re:Fashion | 상품 상세보기"
-          url={`${seoData.url}/shop/${id}`}
-        />
-
-        <Layout noPadding className={$['shop-detail-layout']}>
-          <ProductImgSlide
-            {...{ id, isMe, isSoldOut, imgList: sellerInfo.image }}
-          />
-          <ProfileInfo
-            {...{ userId, profileImage, name, title: basic.title }}
-          />
-
-          <section className={$['shop-detail-info']}>
-            <ProductBasic basic={basic} {...{ id, isMe, isSoldOut }} />
-            <ProductNotice sellerNotice={sellerNotice} />
-            {measure && (
-              <ProductSize size={measure} kind={basic.classification} />
-            )}
-            {opinion && (
-              <SellerComment opinion={opinion} src={sellerInfo.profileImg} />
-            )}
-            <ProductFooter
-              footer={{
-                ...{ price, isIncludeDelivery, updatedAt },
-                ...{ like, view, contact },
-              }}
-            >
-              연락하기
-            </ProductFooter>
-          </section>
-        </Layout>
-      </>
-    );
-  }
-  return null;
 }
+
+ShopDetail.getLayout = function getLayout(page: ReactElement) {
+  return (
+    <Layout noPadding className={$['shop-detail-layout']}>
+      {page}
+    </Layout>
+  );
+};
 
 export default ShopDetail;
