+++ src/pages/shop/[id]/style.module.scss
@@ -1,7 +1,3 @@
 .shop-detail-layout {
   padding-bottom: 100px;
 }
-
-.shop-detail-info {
-  padding: 16px 23px 20px;
-}
